New version of emacs interface from Martin Gregory.
Here's his description of the changes:

"1. it uses gettext to determine the correct version of the "Currently checking"
string and so works for any supported locale;

2. it implements highlighting in the Gramadoir output buffer. Here I have
overcome the <b></b> problem you mentioned by exploiting an Emacs feature called
text properties, the "invisible" property in particular. This means that I can
still find the tags (for the next stage of highlighting in the original text -
to be implemented next) but they are not visible.

3. it gracefully handles a call to next-message if there are no more messages."