Fixes a long-standing minor bug -- with a text like this:
"Féach ar <a href="http://borel.slu.edu/">http://borel.slu.edu/</a>."

the fromx attribute for the misspelling "http" was marked as
the "http" inside the <a>, not the one actually in the text.
This is critical for OpenOffice if one wants to grammar check
the underlying XML format directly and insert markup.